Updated Form and form-adjacent stories
