+++ example/src/stories/Progress/progress.stories.js
@@ -6,7 +6,7 @@ import MyREADME from './README.md';
 
 import { Progress } from 'rimble-ui'
 
-storiesOf('Form', module)
+storiesOf('Form/Inputs', module)
   .addDecorator(withDocs(MyREADME))
   .add('Progress', () => (
     <Progress value={0.5} />
