+++ example/src/stories/Textarea/textarea.stories.js
@@ -6,7 +6,7 @@ import MyREADME from './README.md';
 
 import { Textarea } from 'rimble-ui'
 
-storiesOf('Form', module)
+storiesOf('Form/Inputs', module)
   .addDecorator(withDocs(MyREADME))
   .add('Textarea', () => (
     <Textarea rows={4} defaultValue='Hello' />
